chore: add elastic search sniffing configuration option setting.
Allows users to disable sniffing based on their specific Elasticsearch deplyment instance and environment.
This change is due to issues with proxies and irrelevant hostnames for some elasticsearch instances from cloud providers.
It enables application of sniffing best practices as described in this [link](https://www.elastic.co/blog/elasticsearch-sniffing-best-practices-what-when-why-how)